requirements.txt
aiogram==3.8
mysql-connector-python==8.0
python-dotenv==1.0.0
